Expand to two internal modes, one to configure the local system and the second to control compilation in a standardized way

+++ lib/amakelocal/amakelocal.xpl
@@ -13,9 +13,18 @@ $Program =~ s@.*/@@;
 
 elog_init( $Program, @ARGV );
 
-if( !Getopts( 'ip:v' ) || @ARGV > 1 ) {
+if( !Getopts( 'ip:v' ) ) {
 
-	elog_die( "Usage: amakelocal [-i] [-v] [-p pfname]\n" );
+	elog_die( "Usage: amakelocal [-i] [-v] [-p pfname] [MACRO [MACRO ...]]\n" );
+}
+
+if( @ARGV >= 1 ) {
+
+	$runmode = "verify";
+
+} else {
+
+	$runmode = "construct";
 }
 
 if( $opt_p ) {
@@ -30,14 +39,14 @@ if( $opt_p ) {
 
 if( pffiles( $Pf ) ) {
 
-	if( $opt_v ) {
+	if( $opt_v && $runmode eq "construct" ) {
 		
 		elog_notify( "Using parameter-file '$Pf.pf'\n" );
 	}
 
 } else {
 
-	if( $opt_v ) {
+	if( $opt_v && $runmode eq "construct" ) {
 		
 		elog_notify( "Couldn't find '$Pf.pf'; Using parameter-file '$Pf_proto.pf'\n" );
 	}
@@ -45,42 +54,78 @@ if( pffiles( $Pf ) ) {
 	$Pf = $Pf_proto;
 }
 
-%elements = %{pfget($Pf,"elements")}; 
 $output_file = pfget( $Pf, "output_file" );
-$header = pfget( $Pf, "header" );
 
-open( O, ">$output_file" );
+if( $runmode eq "construct" ) {
 
-print O "$header\n\n";
+	%elements = %{pfget($Pf,"elements")}; 
+	$header = pfget( $Pf, "header" );
 
-foreach $element ( keys( %elements ) ) {
-	
-	if( defined( $elements{$element}{$Os} ) && $elements{$element}{$Os} ne "" ) {
+	open( O, ">$output_file" );
 
-		print O "$element = $elements{$element}{$Os}\n";
+	print O "$header\n\n";
+
+	foreach $element ( keys( %elements ) ) {
+		
+		if( defined( $elements{$element}{$Os} ) && $elements{$element}{$Os} ne "" ) {
+
+			print O "$element = $elements{$element}{$Os}\n";
+		}
 	}
-}
 
-close( O );
-
-if( $opt_v ) {
-	
-	elog_notify( "Generated '$output_file' from parameter-file '$Pf'\n" );
-}
-
-if( $opt_i ) {
+	close( O );
 
 	if( $opt_v ) {
-
-		elog_notify( "Installing '$output_file' in $ENV{ANTELOPE}/include" );
-	} 
-
-	system( "deposit $output_file $ENV{ANTELOPE}/include" );
-
-	unless( cwd() eq "$ENV{ANTELOPE}/include" ) {
-
-		unlink( $output_file );
+		
+		elog_notify( "Generated '$output_file' from parameter-file '$Pf'\n" );
 	}
+
+	if( $opt_i ) {
+
+		if( $opt_v ) {
+
+			elog_notify( "Installing '$output_file' in $ENV{ANTELOPE}/include" );
+		} 
+
+		system( "deposit $output_file $ENV{ANTELOPE}/include" );
+
+		unless( cwd() eq "$ENV{ANTELOPE}/include" ) {
+
+			unlink( $output_file );
+		}
+	}
+
+	exit( 0 );
 }
 
-exit( 0 );
+if( $runmode eq "verify" ) {
+
+	open( A, "$ENV{ANTELOPE}/include/$output_file" );
+
+	@antelopemake = <A>;
+
+	close( A );
+
+	$exitcode = 0;
+
+	foreach $macro ( @ARGV ) {
+	
+		if( ! grep( m/^$macro\s*=/, @antelopemake ) ) {
+
+			$exitcode = 1;
+
+			elog_complain( "Required macro '$macro' is undefined, cancelling " .
+			   "compilation in current subdirectory '" . cwd() . "' " .
+			   "(Use amakelocal(1) to configure your local system).\n" );
+
+		} else {
+
+			if( $opt_v ) {
+
+				elog_complain( "Required macro '$macro' is defined, continuing.\n" );
+			}
+		}
+	}
+
+	exit( $exitcode );
+}
